Extract method about text to the design system module.
